third_party/WebKit/LayoutTests/inspector/console/console-filter-test-expected.txt

CONSOLE MESSAGE: line 9: 0topGroup
CONSOLE MESSAGE: line 3: 1topGroup
CONSOLE MESSAGE: line 9: 2topGroup
CONSOLE MESSAGE: line 3: 3topGroup
CONSOLE MESSAGE: line 9: 4topGroup
CONSOLE MESSAGE: line 3: 5topGroup
CONSOLE MESSAGE: line 9: 6topGroup
CONSOLE MESSAGE: line 3: 7topGroup
CONSOLE MESSAGE: line 9: 8topGroup
CONSOLE MESSAGE: line 3: 9topGroup
CONSOLE MESSAGE: line 22: outerGroup
CONSOLE MESSAGE: line 9: 10outerGroup
CONSOLE MESSAGE: line 3: 11outerGroup
CONSOLE MESSAGE: line 9: 12outerGroup
CONSOLE MESSAGE: line 3: 13outerGroup
CONSOLE MESSAGE: line 9: 14outerGroup
CONSOLE MESSAGE: line 3: 15outerGroup
CONSOLE MESSAGE: line 9: 16outerGroup
CONSOLE MESSAGE: line 3: 17outerGroup
CONSOLE MESSAGE: line 9: 18outerGroup
CONSOLE MESSAGE: line 3: 19outerGroup
CONSOLE MESSAGE: line 29: innerGroup
CONSOLE MESSAGE: line 9: 20innerGroup
CONSOLE MESSAGE: line 3: 21innerGroup
CONSOLE MESSAGE: line 9: 22innerGroup
CONSOLE MESSAGE: line 3: 23innerGroup
CONSOLE MESSAGE: line 9: 24innerGroup
CONSOLE MESSAGE: line 3: 25innerGroup
CONSOLE MESSAGE: line 9: 26innerGroup
CONSOLE MESSAGE: line 3: 27innerGroup
CONSOLE MESSAGE: line 9: 28innerGroup
CONSOLE MESSAGE: line 3: 29innerGroup
CONSOLE MESSAGE: line 36: console.groupEnd
CONSOLE MESSAGE: line 37: console.groupEnd
CONSOLE MESSAGE: line 41: Hello 1
CONSOLE MESSAGE: line 42: Hello 2
CONSOLE MESSAGE: line 44: end
Tests that console can filter messages by source.


Running: beforeFilter
beforeFilter
console-filter-test.html:9 0topGroup
log-source.js:3 1topGroup
console-filter-test.html:9 2topGroup
log-source.js:3 3topGroup
console-filter-test.html:9 4topGroup
log-source.js:3 5topGroup
console-filter-test.html:9 6topGroup
log-source.js:3 7topGroup
console-filter-test.html:9 8topGroup
log-source.js:3 9topGroup
>console-filter-test.html:22 outerGroup
  console-filter-test.html:9 10outerGroup
  log-source.js:3 11outerGroup
  console-filter-test.html:9 12outerGroup
  log-source.js:3 13outerGroup
  console-filter-test.html:9 14outerGroup
  log-source.js:3 15outerGroup
  console-filter-test.html:9 16outerGroup
  log-source.js:3 17outerGroup
  console-filter-test.html:9 18outerGroup
  log-source.js:3 19outerGroup
  >console-filter-test.html:29 innerGroup
    console-filter-test.html:9 20innerGroup
    log-source.js:3 21innerGroup
    console-filter-test.html:9 22innerGroup
    log-source.js:3 23innerGroup
    console-filter-test.html:9 24innerGroup
    log-source.js:3 25innerGroup
    console-filter-test.html:9 26innerGroup
    log-source.js:3 27innerGroup
    console-filter-test.html:9 28innerGroup
    log-source.js:3 29innerGroup
console-filter-test.html:41 Hello 1
console-filter-test.html:42 Hello 2
console-filter-test.html:44 end

Running: addURL1Filter
log-source.js:3 1topGroup
log-source.js:3 3topGroup
log-source.js:3 5topGroup
log-source.js:3 7topGroup
log-source.js:3 9topGroup
>console-filter-test.html:22 outerGroup
  log-source.js:3 11outerGroup
  log-source.js:3 13outerGroup
  log-source.js:3 15outerGroup
  log-source.js:3 17outerGroup
  log-source.js:3 19outerGroup
  >console-filter-test.html:29 innerGroup
    log-source.js:3 21innerGroup
    log-source.js:3 23innerGroup
    log-source.js:3 25innerGroup
    log-source.js:3 27innerGroup
    log-source.js:3 29innerGroup

Running: addURL2Filter
>console-filter-test.html:22 outerGroup
  >console-filter-test.html:29 innerGroup

Running: removeURL1Filter
console-filter-test.html:9 0topGroup
console-filter-test.html:9 2topGroup
console-filter-test.html:9 4topGroup
console-filter-test.html:9 6topGroup
console-filter-test.html:9 8topGroup
>console-filter-test.html:22 outerGroup
  console-filter-test.html:9 10outerGroup
  console-filter-test.html:9 12outerGroup
  console-filter-test.html:9 14outerGroup
  console-filter-test.html:9 16outerGroup
  console-filter-test.html:9 18outerGroup
  >console-filter-test.html:29 innerGroup
    console-filter-test.html:9 20innerGroup
    console-filter-test.html:9 22innerGroup
    console-filter-test.html:9 24innerGroup
    console-filter-test.html:9 26innerGroup
    console-filter-test.html:9 28innerGroup
console-filter-test.html:41 Hello 1
console-filter-test.html:42 Hello 2
console-filter-test.html:44 end

Running: restoreURL1Filter
>console-filter-test.html:22 outerGroup
  >console-filter-test.html:29 innerGroup

Running: removeAllFilters
console-filter-test.html:9 0topGroup
log-source.js:3 1topGroup
console-filter-test.html:9 2topGroup
log-source.js:3 3topGroup
console-filter-test.html:9 4topGroup
log-source.js:3 5topGroup
console-filter-test.html:9 6topGroup
log-source.js:3 7topGroup
console-filter-test.html:9 8topGroup
log-source.js:3 9topGroup
>console-filter-test.html:22 outerGroup
  console-filter-test.html:9 10outerGroup
  log-source.js:3 11outerGroup
  console-filter-test.html:9 12outerGroup
  log-source.js:3 13outerGroup
  console-filter-test.html:9 14outerGroup
  log-source.js:3 15outerGroup
  console-filter-test.html:9 16outerGroup
  log-source.js:3 17outerGroup
  console-filter-test.html:9 18outerGroup
  log-source.js:3 19outerGroup
  >console-filter-test.html:29 innerGroup
    console-filter-test.html:9 20innerGroup
    log-source.js:3 21innerGroup
    console-filter-test.html:9 22innerGroup
    log-source.js:3 23innerGroup
    console-filter-test.html:9 24innerGroup
    log-source.js:3 25innerGroup
    console-filter-test.html:9 26innerGroup
    log-source.js:3 27innerGroup
    console-filter-test.html:9 28innerGroup
    log-source.js:3 29innerGroup
console-filter-test.html:41 Hello 1
console-filter-test.html:42 Hello 2
console-filter-test.html:44 end

Running: checkContextFilter
>console-filter-test.html:22 outerGroup
  >console-filter-test.html:29 innerGroup
console-filter-test.html:41 Hello 1

Running: checkAllContextsFilter
console-filter-test.html:9 0topGroup
log-source.js:3 1topGroup
console-filter-test.html:9 2topGroup
log-source.js:3 3topGroup
console-filter-test.html:9 4topGroup
log-source.js:3 5topGroup
console-filter-test.html:9 6topGroup
log-source.js:3 7topGroup
console-filter-test.html:9 8topGroup
log-source.js:3 9topGroup
>console-filter-test.html:22 outerGroup
  console-filter-test.html:9 10outerGroup
  log-source.js:3 11outerGroup
  console-filter-test.html:9 12outerGroup
  log-source.js:3 13outerGroup
  console-filter-test.html:9 14outerGroup
  log-source.js:3 15outerGroup
  console-filter-test.html:9 16outerGroup
  log-source.js:3 17outerGroup
  console-filter-test.html:9 18outerGroup
  log-source.js:3 19outerGroup
  >console-filter-test.html:29 innerGroup
    console-filter-test.html:9 20innerGroup
    log-source.js:3 21innerGroup
    console-filter-test.html:9 22innerGroup
    log-source.js:3 23innerGroup
    console-filter-test.html:9 24innerGroup
    log-source.js:3 25innerGroup
    console-filter-test.html:9 26innerGroup
    log-source.js:3 27innerGroup
    console-filter-test.html:9 28innerGroup
    log-source.js:3 29innerGroup
console-filter-test.html:41 Hello 1
console-filter-test.html:42 Hello 2
console-filter-test.html:44 end

Running: checkTextFilter
>console-filter-test.html:22 outerGroup
  console-filter-test.html:9 10outerGroup
  log-source.js:3 11outerGroup
  console-filter-test.html:9 12outerGroup
  log-source.js:3 13outerGroup
  console-filter-test.html:9 14outerGroup
  log-source.js:3 15outerGroup
  console-filter-test.html:9 16outerGroup
  log-source.js:3 17outerGroup
  console-filter-test.html:9 18outerGroup
  log-source.js:3 19outerGroup
  >console-filter-test.html:29 innerGroup

Running: checkResetFilter
console-filter-test.html:9 0topGroup
log-source.js:3 1topGroup
console-filter-test.html:9 2topGroup
log-source.js:3 3topGroup
console-filter-test.html:9 4topGroup
log-source.js:3 5topGroup
console-filter-test.html:9 6topGroup
log-source.js:3 7topGroup
console-filter-test.html:9 8topGroup
log-source.js:3 9topGroup
>console-filter-test.html:22 outerGroup
  console-filter-test.html:9 10outerGroup
  log-source.js:3 11outerGroup
  console-filter-test.html:9 12outerGroup
  log-source.js:3 13outerGroup
  console-filter-test.html:9 14outerGroup
  log-source.js:3 15outerGroup
  console-filter-test.html:9 16outerGroup
  log-source.js:3 17outerGroup
  console-filter-test.html:9 18outerGroup
  log-source.js:3 19outerGroup
  >console-filter-test.html:29 innerGroup
    console-filter-test.html:9 20innerGroup
    log-source.js:3 21innerGroup
    console-filter-test.html:9 22innerGroup
    log-source.js:3 23innerGroup
    console-filter-test.html:9 24innerGroup
    log-source.js:3 25innerGroup
    console-filter-test.html:9 26innerGroup
    log-source.js:3 27innerGroup
    console-filter-test.html:9 28innerGroup
    log-source.js:3 29innerGroup
console-filter-test.html:41 Hello 1
console-filter-test.html:42 Hello 2
console-filter-test.html:44 end
